requirements.txt
openai
jsonlines
SPARQLWrapper
argparse
python-dotenv
config
vllm
pandas